refactoring to appropriate package
